Finalizing permissions

+++ spatie-app/routes/api.php
@@ -3,6 +3,8 @@
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\TaskController;
 use App\Http\Controllers\DepartmentController;
+use App\Http\Controllers\UserController;
+use App\Models\User;
 
 /*
 |--------------------------------------------------------------------------
@@ -15,56 +17,50 @@ use App\Http\Controllers\DepartmentController;
 |
 */
 
-Route::post('/create', [TaskController::class, 'assignTask']);
-Route::post('/show', [TaskController::class, 'show']);
-
-Route::delete('/delete/{id}', [TaskController::class, 'destroy']);
-
-Route::put('/update/{id}', [TaskController::class, 'Update_task']); // for updating complete start
-Route::put('/reassign/{id}', [TaskController::class, 'reassign_task']); // for reassigning the task
-//Route::resource('teams', TeamController::class);
-
-// '/api/team-members' -> this endpoint is for TeamMember
-
-//Route::resource('teammembers', TeamMemberController::class);
-
-
-// Routes for TeamController
-Route::get('/teamindex', [TeamController::class, 'index']);
-
-Route::post('/teamcreate', [TeamController::class, 'store']);
-
-Route::get('/teamshow/{id}', [TeamController::class, 'show']);
-
-Route::put('/teamupdate', [TeamController::class, 'update']);
-
-Route::delete('/teamdestroy/{id}', [TeamController::class, 'destroy']);
-
-// Routes for TeamMemberController
-Route::get('/memberindex', [TeamMemberController::class, 'index']);
-
-Route::post('/membercreate', [TeamMemberController::class, 'store']);
-
-Route::get('/membershow/{id}', [TeamMemberController::class, 'show']);
-
-Route::put('/memberupdate/{id}', [TeamMemberController::class, 'update']);
-
-Route::delete('/memberdestroy/{id}', [TeamMemberController::class, 'destroy']);
-
-
 
 Route::post('/login', [UserController::class, 'login']);
 Route::post('/register', [UserController::class, 'register']);
-Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
-    return $request->user();
+
+
+
+
+Route::middleware('auth:sanctum')->group(function(){
+    Route::get('/users', [UserController::class, 'getUsers'])->middleware('permission:can-access-all-users');
+    Route::post('/create', [TaskController::class, 'assignTask'])->middleware('permission:can-create-task');
+    Route::post('/show', [TaskController::class, 'show'])->middleware('permission:can-view-task');
+    Route::delete('/delete/{id}', [TaskController::class, 'destroy'])->middleware('permission:can-delete-task');
+    Route::put('/update/{id}', [TaskController::class, 'Update_task'])->middleware('permission:can-update-task'); // for updating complete start
+    Route::put('/reassign/{id}', [TaskController::class, 'reassign_task'])->middleware('permission:can-reassign-task'); // for reassigning the task
 });
 
-Route::post('add',[DepartmentController::class,'adding']);
 
-Route::put('update/{id}',[DepartmentController::class,'updating']);
+// Routes for TeamController
+Route::middleware('auth:sanctum')->group(function(){
+Route::get('/teamindex', [TeamController::class, 'index'])->middleware('permission:can-view-teams');
+Route::post('/teamcreate', [TeamController::class, 'store'])->middleware('permission:can-create-teams');
+Route::get('/teamshow/{id}', [TeamController::class, 'show'])->middleware('permission:can-view-specific-team');
+Route::put('/teamupdate', [TeamController::class, 'update'])->middleware('permission:can-update-teams');
+Route::delete('/teamdestroy/{id}', [TeamController::class, 'destroy'])->middleware('permission:can-delete-team');
+});
 
-Route::delete('delete/{id}',[DepartmentController::class,'delete']);
+// Routes for TeamMemberController
+Route::middleware('auth:sanctum')->group(function(){
+Route::get('/memberindex', [TeamMemberController::class, 'index'])->middleware('permission:can-view-members');
+Route::post('/membercreate', [TeamMemberController::class, 'store'])->middleware('permission:can-create-members');
+Route::get('/membershow/{id}', [TeamMemberController::class, 'show'])->middleware('permission:can-view-specific-member');
+Route::put('/memberupdate/{id}', [TeamMemberController::class, 'update'])->middleware('permission:can-update-member');
+Route::delete('/memberdestroy/{id}', [TeamMemberController::class, 'destroy'])->middleware('permission:can-delete-member');
+});
 
-Route::get('get',[DepartmentController::class,'getData']);
+
+
+
+Route::middleware()->group(function(){
+    Route::post('add',[DepartmentController::class,'adding'])->middleware('permission:can-add-department');
+    Route::put('update/{id}',[DepartmentController::class,'updating'])->middleware('permission:can-update-department');
+    Route::delete('delete/{id}',[DepartmentController::class,'delete'])->middleware('permission:can-delete-department');
+    Route::get('get',[DepartmentController::class,'getData'])->middleware('permission:can-view-department');
+});
+
